% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prio.R
\name{comb}
\alias{comb}
\title{Strain combination builder}
\usage{
comb(geno, min_strain_benef = 0.1, max_set_size = 3)
}
\arguments{
\item{geno}{Data frame of genotypes for additional strains.}

\item{min_strain_benef}{Minimum reduction factor (min) of a single strain.
Default is 0.1.}

\item{max_set_size}{Maximum set of strains. Default is 3.}
}
\value{
Data frame
}
\description{
Generate strain sets and calculate reduction factors
}
\keyword{internal}
