% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{df_split}
\alias{df_split}
\title{Splits data frame df into subsets with maximum n rows}
\usage{
df_split(df, n)
}
\arguments{
\item{df}{Data frame.}

\item{n}{Max number of rows per subset.}
}
\value{
List of data frames.
}
\description{
Splits data frame df into subsets with maximum n rows
}
\keyword{internal}
