% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ensembl_rest_vep}
\alias{ensembl_rest_vep}
\title{Request variant consequences from Variant Effect Predictor (VEP) via Ensembl Rest Service}
\usage{
ensembl_rest_vep(snps, species)
}
\arguments{
\item{snps}{Data frame including columns rsid, ref, alt.}

\item{species}{Species name, e.g. mouse or human.}
}
\value{
Data frame.
}
\description{
Request variant consequences from Variant Effect Predictor (VEP) via Ensembl Rest Service
}
\keyword{internal}
