% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{filter_query}
\alias{filter_query}
\title{Filter query builder}
\usage{
filter_query(chr, start = NA, end = NA, consequence = NA, impact = NA)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position. NA by default.}

\item{end}{Optional chromosomal end position. NA by default.}

\item{consequences}{Vector containing consequence types. NA by default.}

\item{impacts}{Vector containing impact types. NA by default.}
}
\value{
Query string.
}
\description{
Filter query builder
}
\keyword{internal}
