% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmusserv.R
\name{genehopper_request}
\alias{genehopper_request}
\title{Send HTTP request to MMUS Server}
\usage{
genehopper_request(q, n.tries = 2, method = "GET")
}
\arguments{
\item{q}{Query string}

\item{n.tries}{Number of tries}

\item{method}{HTTP method to use}
}
\value{
Data frame.
}
\description{
Send HTTP request to MMUS Server
}
\examples{
genehopper_request("http://mmusserv.genehopper.de/rest/mmusprio/chr1:5000000-6000000")
}
\keyword{internal}
