% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prio.R
\name{get_top}
\alias{get_top}
\title{Best strain combinations}
\usage{
get_top(red, n_top)
}
\arguments{
\item{red}{Reduction factors data frame.}

\item{n_top}{Number of combinations to be returned.}
}
\value{
Data frame
}
\description{
Get best strain combinations
}
\examples{
l = prio("chr1",
  start = 5000000, end = 6000000,
  strain1 = "C57BL_6J", strain2 = "AKR_J"
)

get_top(l$reduction, 3)
}
