% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{getURL}
\alias{getURL}
\title{Get backend service url}
\usage{
getURL()
}
\value{
URL string.
}
\description{
Get backend service URL. Default: http://45.85.146.64:9000/rest/finemap/
}
\examples{
getURL()
}
