% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_fetch.R
\name{mmusfetch}
\alias{mmusfetch}
\title{MMUS Fetch}
\usage{
mmusfetch(
  chr,
  start = NA,
  end = NA,
  consequence = NA,
  impact = NA,
  return_obj = "dataframe"
)
}
\arguments{
\item{chr}{Chromosome name or vector of chromosome names.}

\item{start}{Optional Chromosomal start position of target region (GRCm38). Multiple positions can be passed as vector. NA by default.}

\item{end}{Optional Chromosomal end position of target region (GRCm38). Multiple positions can be passed as vector. NA by default.}

\item{consequence}{Vector containing consequence types. NA by default.}

\item{impact}{Vector containing impact types. NA by default.}

\item{return_obj}{The user can choose to get the result to be returned
as data frame ("dataframe") or as a GenomicRanges::GRanges ("granges") object. Default value is "dataframe".}
}
\value{
Data frame or GenomicRanges::GRanges object containing result data.
}
\description{
Detch genetic variant data for chromosomal region
}
\examples{
mmusfetch("chr7", start=5000000, end=6000000)
}
