% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_finemap.R
\name{mmusfinemap_query}
\alias{mmusfinemap_query}
\title{URL builder}
\usage{
mmusfinemap_query(
  chr,
  start = NA,
  end = NA,
  strain1,
  strain2,
  consequence = NA,
  impact = NA,
  thr1 = 0,
  thr2 = 0
)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position. NA by default.}

\item{end}{Optional chromosomal end position. NA by default.}

\item{thr1}{Number discordant strains in strain1. Between 0 and length(strain1)-1. 0 by default.}

\item{thr2}{Number discordant strains in strain2. Between 0 and length(strain2)-1. 0 by default.}

\item{consequences}{Vector containing consequence types. NA by default.}

\item{impacts}{Vector containing impact types. NA by default.}
}
\value{
Query string.
}
\description{
URL builder
}
\examples{
mmusfinemap_query("chr1", start=5000000, end=6000000, strain1=c("C57BL_6J"), strain2=c("129S1_SvImJ", "129S5SvEvBrd", "AKR_J"), impact=c("HIGH", "MODERATE"), consequence=c("feature_truncation", "regulatory_region_variant"))
mmusfinemap_query("chr1", start=5000000, end=6000000, strain1=c("C57BL_6J"), strain2=c("AKR_J", "A_J", "BALB_cJ"))
}
\keyword{internal}
