% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{mmusprio}
\alias{mmusprio}
\title{MMUS Prio}
\usage{
mmusprio(
  chr,
  start = NA,
  end = NA,
  consequence = NA,
  impact = NA,
  return_obj = "dataframe"
)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position of target region (GRCm38). NA by default.}

\item{end}{Optional chromosomal end position of target region (GRCm38). NA by default.}

\item{consequence}{Vector containing consequence types. NA by default.}

\item{impact}{Vector containing impact types. NA by default.}

\item{return_obj}{The user can choose to get the result to be returned
as data frame ("dataframe") or as a GenomicRanges::GRanges ("granges") object. Default value is "dataframe".}
}
\value{
Data frame or GenomicRanges::GRanges object containing result data.
}
\description{
Prioritize mouse strains
}
\examples{
mmusprio("chr1", start=5000000, end=6000000)
}
