% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{mmusprio_query}
\alias{mmusprio_query}
\title{URL builder}
\usage{
mmusprio_query(chr, start = NA, end = NA, consequence = NA, impact = NA)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position. NA by default.}

\item{end}{Optional chromosomal end position. NA by default.}

\item{consequences}{Vector containing consequence types. NA by default.}

\item{impacts}{Vector containing impact types. NA by default.}
}
\value{
Query string.
}
\description{
URL builder
}
\examples{
mmusprio_query("chr1", start=5000000, end=6000000, impact=c("HIGH", "MODERATE"), consequence=c("feature_truncation", "regulatory_region_variant"))
}
\keyword{internal}
