% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prio.R
\name{reduction}
\alias{reduction}
\title{Reduction factor calculation}
\usage{
reduction(combs, geno)
}
\arguments{
\item{combs}{Data frame of strain sets.}

\item{geno}{Data frame of genotypes for additional strains.}
}
\value{
Dataframe
}
\description{
Generate strain sets and calculate reduction factors
}
\keyword{internal}
