% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{setURL}
\alias{setURL}
\title{Set backend service url}
\usage{
setURL(url)
}
\arguments{
\item{url}{URL of backend service.}
}
\value{
No return value.
}
\description{
Set backend service URL. Default: http://mousefm.genehopper.de/rest/finemap/
}
\examples{
setURL("http://backendserver.com")
}
