% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_consequences}
\alias{annotate_consequences}
\title{Annotate consequences}
\usage{
annotate_consequences(snps, species)
}
\arguments{
\item{snps}{Data frame including columns rsid, ref, alt.}

\item{species}{Species name, e.g. mouse (GRCm38) or human (GRCh38).}
}
\value{
Data frame.
}
\description{
Request variant consequences from Variant Effect Predictor (VEP) via Ensembl Rest Service.
}
\examples{
geno = mmusfinemap("chr1", start=5000000, end=6000000,
strain1=c("C57BL_6J"), strain2=c("AKR_J", "A_J", "BALB_cJ"))

df = annotate_consequences(geno, "mouse")
}
