% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{comb}
\alias{comb}
\title{Strain combination builder}
\usage{
comb(geno, min_strain_benef = 0.1, max_set_size = 3)
}
\arguments{
\item{geno}{Data frame of genotypes for additional strains.}

\item{min_strain_benef}{Minimum reduction factor (min) of a single strain.}

\item{max_set_size}{Maximum set of strains.}
}
\value{
Dataframe
}
\description{
Generate strain sets and calculate reduction factors
}
\keyword{internal}
