% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{filter_query}
\alias{filter_query}
\title{Filter query builder}
\usage{
filter_query(chr, start = NULL, end = NULL, consequence = NULL, impact = NULL)
}
\arguments{
\item{chr}{Vector of chromosome names.}

\item{start}{Optional vector of chromosomal start positions of target regions (GRCm38).}

\item{end}{Optional vector of chromosomal end positions of target regions (GRCm38).}

\item{consequence}{Optional vector of consequence types.}

\item{impact}{Optional vector of impact types.}
}
\value{
Query string.
}
\description{
Filter query builder
}
\keyword{internal}
