% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{finemap_query}
\alias{finemap_query}
\title{Finemap query builder}
\usage{
finemap_query(
  chr,
  start = NA,
  end = NA,
  strain1,
  strain2,
  consequence = NA,
  impact = NA,
  thr1 = 0,
  thr2 = 0
)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position. NA by default.}

\item{end}{Optional chromosomal end position. NA by default.}

\item{strain1}{First strain(s).}

\item{strain2}{Second strain(s).}

\item{thr1}{Number discordant strains in strain1. Between 0 and length(strain1)-1. 0 by default.}

\item{thr2}{Number discordant strains in strain2. Between 0 and length(strain2)-1. 0 by default.}

\item{consequences}{Vector containing consequence types. NA by default.}

\item{impacts}{Vector containing impact types. NA by default.}
}
\value{
Query string.
}
\description{
Finemap query builder
}
\keyword{internal}
