% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{get_top}
\alias{get_top}
\title{Best combinations}
\usage{
get_top(rf, n_top)
}
\arguments{
\item{rf}{Reduction factors data frame.}

\item{n_top}{Number if combinations to be returned.}
}
\value{
Dataframe
}
\description{
Get best strain combinations
}
\examples{
l = mmusprio("chr1", start=5000000, end=6000000, strain1="C57BL_6J", strain2="AKR_J")

get_top(l$reduction, 3)
}
