% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_finemap.R
\name{mmusfinemap}
\alias{mmusfinemap}
\title{MMUS Finemapping}
\usage{
mmusfinemap(
  chr,
  start = NA,
  end = NA,
  strain1,
  strain2,
  consequence = NA,
  impact = NA,
  thr1 = 0,
  thr2 = 0,
  return_obj = "dataframe"
)
}
\arguments{
\item{chr}{Chromosome name.}

\item{start}{Optional chromosomal start position of target region (GRCm38). NA by default.}

\item{end}{Optional chromosomal end position of target region (GRCm38). NA by default.}

\item{strain1}{First strain(s).}

\item{strain2}{Second strain(s).}

\item{consequence}{Vector containing consequence types. NA by default.}

\item{impact}{Vector containing impact types. NA by default.}

\item{thr1}{Number discordant strains in strain1. Between 0 and length(strain1)-1. 0 by default.}

\item{thr2}{Number discordant strains in strain2. Between 0 and length(strain2)-1. 0 by default.}

\item{return_obj}{The user can choose to get the result to be returned
as data frame ("dataframe") or as a GenomicRanges::GRanges ("granges") object. Default value is "dataframe".}
}
\value{
Data frame or GenomicRanges::GRanges object containing result data.
}
\description{
Finemapping
}
\examples{
mmusfinemap("chr1", start=5000000, end=6000000,
strain1=c("C57BL_6J"), strain2=c("129S1_SvImJ", "129S5SvEvBrd", "AKR_J"))

mmusfinemap("chr1", start=5000000, end=6000000,
strain1=c("C57BL_6J"), strain2=c("AKR_J", "A_J", "BALB_cJ"))

mmusfinemap("chr1", start=5000000, end=6000000,
strain1=c("C57BL_6J"), strain2=c("NOD_ShiLtJ", "CBA_J", "KK_HiJ", "PWK_PhJ"), thr2=1)
}
