% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{mmusprio}
\alias{mmusprio}
\title{MMUS Prio}
\usage{
mmusprio(
  chr,
  start = NULL,
  end = NULL,
  strain1,
  strain2,
  consequence = NULL,
  impact = NULL,
  min_strain_benef = 0.1,
  max_set_size = 3
)
}
\arguments{
\item{chr}{Vector of chromosome names.}

\item{start}{Optional vector of chromosomal start positions of target regions (GRCm38).}

\item{end}{Optional vector of chromosomal end positions of target regions (GRCm38).}

\item{strain1}{First strain.}

\item{strain2}{Second strain.}

\item{consequence}{Optional vector of consequence types.}

\item{impact}{Optional vector of impact types.}

\item{min_strain_benef}{Minimum reduction factor (min) of a single strain.}

\item{max_set_size}{Maximum set of strains.}
}
\value{
Dataframe
}
\description{
Prioritize mouse strains
}
\examples{
res = mmusprio("chr1", start=5000000, end=6000000, strain1="C57BL_6J", strain2="AKR_J")

comment(res$genotypes)
}
