% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{mmusprio}
\alias{mmusprio}
\title{MMUS Prio}
\usage{
mmusprio(
  chr,
  start = NA,
  end = NA,
  strain1,
  strain2,
  consequence = NA,
  impact = NA,
  min_strain_benef = 0.1,
  max_set_size = 3
)
}
\arguments{
\item{chr}{Chromosome name or vector of chromosome names.}

\item{start}{Optional Chromosomal start position of target region (GRCm38). Multiple positions can be passed as vector. NA by default.}

\item{end}{Optional Chromosomal end position of target region (GRCm38). Multiple positions can be passed as vector. NA by default.}

\item{strain1}{First strain.}

\item{strain2}{Second strain.}

\item{consequence}{Vector containing consequence types. NA by default.}

\item{impact}{Vector containing impact types. NA by default.}

\item{min_strain_benef}{Minimum reduction factor (min) of a single strain.}

\item{max_set_size}{Maximum set of strains.}
}
\value{
Dataframe
}
\description{
Prioritize mouse strains
}
\examples{
res = mmusprio("chr1", start=5000000, end=6000000, strain1="C57BL_6J", strain2="AKR_J")

comment(res$genotypes)
}
