% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmus_prio.R
\name{vis_reduction_factors}
\alias{vis_reduction_factors}
\title{Visualize}
\usage{
vis_reduction_factors(geno, rf, n_top)
}
\arguments{
\item{geno}{Genotype data frame.}

\item{rf}{Reduction factor data frame.}

\item{n_top}{Number if combinations to be returned.}
}
\value{
Dataframe
}
\description{
Visualize reduction factors
}
\examples{
r = mmusprio("chr1", start=5000000, end=6000000, strain1="C57BL_6J", strain2="AKR_J")
plots = vis_reduction_factors(r$genotypes, r$reduction, 2)
plots[[1]]
plots[[2]]
}
